requirements.txt
torch==2.3.0
transformers==4.41.1
diffusers==0.27.2
huggingface_hub==0.23.1
accelerate==0.31.0
safetensors==0.4.2
python-dotenv==1.0.1
Pillow>=9.0.0